Split message e2e specs and cover auth branches
